Move UserFactory class into the user folder under entity
