+++ src/test/java/com/yi/online_bookshop/service/UserServiceTest.java
@@ -18,8 +18,8 @@ import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.dao.DataIntegrityViolationException;
 
-import com.yi.online_bookshop.domain.entity.UserFactory;
 import com.yi.online_bookshop.domain.entity.user.User;
+import com.yi.online_bookshop.domain.entity.user.UserFactory;
 import com.yi.online_bookshop.dto.user.CreateUserDTO;
 import com.yi.online_bookshop.repository.UserRepository;
